backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.0.0
python-decouple==3.8
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
django-environ==0.11.2

